[SPH] move node rendering nodes to standalone files
