+++ src/shammodels/sph/CMakeLists.txt
@@ -63,6 +63,9 @@ set(Sources
     src/modules/self_gravity/SGSFMMPlummer.cpp
     src/modules/render/CartesianRender.cpp
     src/modules/render/RenderFieldGetter.cpp
+    src/modules/render/SPHInterpolation.cpp
+    src/modules/render/SPHColumnInteg.cpp
+    src/modules/render/SPHAzymuthalInteg.cpp
     src/modules/io/VTKDump.cpp
 )
